Move initial table sizing to the right place in HHBBC
Summary:
HHBBC (and by extension, it's extern-worker jobs) is supposed
to size the static string table and named entity tables according to
the value in RepoGlobalData. This is populated with the configured
value when HPHPc runs.

However it was loading the values at the wrong time. It needs to be
set before RuntimeOptions::Load(), because that's what triggers the
tables to resize. Move the setting to the right place.

This isn't a problem for HHBBC usually because it's usually run as
part of HPHPc, so the runtime options are already set correctly. It
is, however, a problem for extern-worker jobs. This makes the jobs run
slower because the static string table is too small.

Reviewed By: oulgen

Differential Revision: D41010385

fbshipit-source-id: 3fd18becbdb29e2075b87a6ef87ee63876d9ad5e