Enable interprocedural constraints to propagate from previous passes
Summary:
## What

Before this change:
- HIPS used the constraint map from round 0 in later phases, ignoring the latest constraints when doing substitution.

After this change:
- HIPS gives Shape_analysis the latest constraints for each function.

## Behavior change (more coming in the next diff)

The existing recursion.php snapshot test is a great example of the difference

Before, the 'SK a => mixed' propagated from `f` to `main`
But the similar 'SK b => mixed' did **NOT** propagate from `g` to `f` to `main`. With this change, both (unfortunate!) validations propagate.

```
function f(dict<string, mixed> $d) : void {
  $d['a'];
  g($d);
}

function g(dict<string, mixed> $d) : void {
  $d['b'];
  f($d);
}

function main(): void {
  // before: inferred `shape('a' => mixed, 'b' => int)`
  // before: inferred `shape('a' => mixed, 'b' => mixed)`
  f(dict['a' => true, 'b' => 42]);
}
```

Of course that combining the information "a key exists" and "the type for the key is `int`" gives us "the type of for the key is `mixed`" is undesirable, but to me that seems like a separate problem with how we combine information in shape analysis, not how the information is propagated.

## Why this change

I can't come up with a principled reason for the existing behavior of special-casing first-round constraints.

More importantly, this change enables us to handle `return`. Operationally:
- We already have a subset constraint between a return expression and the type hint for the function in which it appears
- D40807745 we propagate constraints from function-return-types to where the function is called. We do not change the existing behavior that only *key* constraints are propagated (a subset constraint that points to an entity in another procedure would be of no use to Shape_analysis, which looks at only one procedure at a time)
- Shape_analysis_solver internally uses subset constraints to propagate key constraints
- With the change in this diff, these new key constraints (from the previous bullet) can be propagated

Reviewed By: madgen

Differential Revision: D40807749

fbshipit-source-id: 6cd9fc0756948165b69932a45d8e8076c69096dd