Add lint and autofix for redundant UNSAFE_CASTs
Summary:
An application of `UNSAFE_CAST` is redundant when the type of the expression being cast is a subtype of the type we are casting to.

This diff adds a lint to warn against such usages along with an autofix to remove the `UNSAFE_CAST` call.

The lint will not fire when the expression has type `Tany` since that indicates there is no type for the expression but will satisfy the `is_subtype` predicate; this condition is also necessary to maintain the behaviour in `unsafe_cast_nondenotable_lower_bound2.php`

Reviewed By: andrewjkennedy

Differential Revision: D39971230

fbshipit-source-id: 02d4f4b018aecbaa49c4250e3297eb2447bd8e3c