Add Two More Methods - Sync and Type or Type Alias Path
Summary:
There are use cases where we just want the path for a type or a type alias and
we don't care which it is.  Currently the API requires calling one method and
then the other to figure that out.  This provides a method which avoids having
to make two calls.

Also, it's handy to have a method which can block until facts is up to date.
This won't guarantee that facts is up to date as of the return of the call,
since another change could have happened, but it at least guarantees it's up to
date as of the beginning of the call.

Reviewed By: edwinsmith

Differential Revision: D40911046

fbshipit-source-id: 50ae0bf9b240b121011715d878f68d9d9064b6ce