Add tests for validation of `void` and `notreturn` type hints
Summary:
Our `naming` phase will raise an error and subsitute with `Herr` when `void` or `noreturn` are used as toplevel hints somewhere other than in return position. When used in return position or as a type parameter of another type in argument position, we should allow.

The diff adds tests which exercise this logic.

Reviewed By: andrewjkennedy

Differential Revision: D40756254

fbshipit-source-id: 36671688115f0cc2dd52258637256f1723abde37