fix bad todo code
Summary:
For TODO functions instead of making a fake CGetL with a bad variable let's make a dummy call and pass the reason as a string parameter.

This fixes textual so instead of outputting this:
```
n0: *Mixed = load &Array(Keyset(K...tBytesId(17))})))
```

We get this:
```
n0: *Mixed = call todo(null, "Array(K...itBytesId(17))})))")
```

While in there fixed the constant generation code to not stomp on itself.

Reviewed By: artempyanykh

Differential Revision: D41007063

fbshipit-source-id: cb3647535d3d214d7d7f25b70b19a8b846fef2c8