Refactor out sketchy Symbolizer header dep on EvictingCacheMap
Summary:
For example, folly/singleton non-export depends on symbolizer, and
symbolizer export-depends on evicting_cache_map. Unfortunately, adding more
dependencies to evicting_cache_map can break users of singleton because they
aren't in the exported deps. But it would be a bad idea to make singleton or
symbolizer heavy on exported deps, so it's easy enough to hide the SymbolCache
behind a unique_ptr. Symbolizer should not be so performance sensitive here to worry about indirection overhead.

Reviewed By: Gownta

Differential Revision: D40915701

fbshipit-source-id: 651678d8c1776c2b0b0ef5300298e1169fa73422