Demote option mixed/null error to a lint
Summary:
This error is not really an error as both the static and dynamic semantics can handle it well. It is only a redundancy. It breaks the substitution property and is non-monotonic (specialising a generic to `?mixed` and `?null` leads to an error.)

This diff demotes the error to a lint warning.

Reviewed By: vsiles

Differential Revision: D41060133

fbshipit-source-id: 935fa22eeb19334c510972afc7a1ea458ad5973e