Interpret intersections as enforced
Summary:
Explicit pessimisation introduces intersections such as
```
<<__SupportDynamicType>>
function foo():(~E & int) { ... }
```
For the purposes of checking SDT functions, we should assume that intersections are enforced component-wise. Also, if the "enforced" return type matches the Hack return type (in fact, it's enough to be a subtype), then we can treat the return type as "fully enforced" and allow dynamic to flow to it by adding `~` to the return type as it is used for checking `return` statements. For example, we should accept the following:
```
<<__SupportDynamicType>>
function foo():(~E & int) {
  return getLikeE();
}
function getLikeE():~E { ... }
```
This is simply analogous to allowing a value of type `~int` to be returned from an SDT function whose return type is declared to be `int`.

Also add a couple of fixes to enforcement: do no treat a generic parameter bound on another generic parameter as enforced (this matches HHVM), and, for now, assume that generic classes with type arguments aren't enforced at all.

Differential Revision: D41006507

fbshipit-source-id: dfe4ec51a9f9e9344fe95f7a2686636cc4854e16