Optimize HPHP::CacheManager::existsHelper by switching to Folly containers
Summary:
Replaced the std::map containing all file paths in sorted order with a folly::F14FastMap. All file lookups should now be O(1). This can save up to 0.001% of our CPU budget, which is meaningful at scale.

Additional changes were required to support directory listing (i.e. CacheManager::readDirectory()). The existing code relied on the sorted map, where all children (and grandchildren etc.) were sorted immediately after the entry for the directory itself. This could not work with the unordered hash table, so the immediate children of a directory are now stored on the CacheData for the directory. This may have an incremental performance boost as well: Instead of an O(N lg N) traversal of the sorted map (where N is the total number of files anywhere under a directory), directory listing is now a simple copy of a vector of size M (where M is only the number of immediate children). This may help slightly in situations like D1943184 (https://github.com/facebook/hhvm/commit/6b39a4e99e9ff55f0893dc6b330416824dcac781).

Finally, in order to not disrupt the child-list on directories, the code to cache all parent directories of a file now runs tip->root with an early out as soon as a parent is found. The old code ran root->tip rebuilding everything unconditionally, so populating a large cache will now involve slightly less thrashing.

Reviewed By: ricklavoie

Differential Revision: D40734705

fbshipit-source-id: d9acaac16f4f30c9122e5263164e57ab42c4117f