Don't treat Closure as a resolved class
Summary:
Closure (the base class for all closures) can be very
pathological. The subclass list will have an entry for every closure
in the code base. For large code bases this list can grow quite
large. This uses a lot of memory and can be slow to
calculate. Moreover, it's not terribly useful. Almost all of the
ClassInfo properties for Closure are fixed due to invariants on
closures. We also rarely have a situation where we know something is a
Closure, but not a specific closure.

So, add special logic to force Closure to be unresolved. It is now an
invariant that a res::Class can never contain a resolved Closure. This
does not affect any relationship checks on Closure due to the previous
work to allow for unresolved and resolved classes to mix.

In a similar vein, don't calculate a FuncFamily for `__invoke`. Like
Closure itself, this can be pathologically large and is not useful. We
almost never have a situation where we're calling `__invoke` on
something completely unknown. Furthermore, on any code base of any
size, the return type is going to invariably be InitCell anyways.

Reviewed By: edwinsmith

Differential Revision: D41069596

fbshipit-source-id: e9bceba587b703eb288d144189d1d319fe73a9c9