Allow for unresolved and resolved res::Class to be each other in HHBBC
Summary:
Right now it's assumed that unresolved and resolved res::Class are
always disjoint. That is, we know, a-priori, that an unresolved
res::Class cannot be a resolved res::Class and vice versa. This has
been sufficient for now because the only way we can get an unresolved
class is if the class doesn't exist.

However, I want to add support for selectively not building (large and
expensive) ClassInfo data for some classes (namely Closure). The
easiest way to support this (right now anyways) is to simply force
such classes to be unresolved. Once we do this, they are no longer
disjoint! We might have a specific closure (which will be resolved),
which "could-be" Closure.

So, we need to improve this logic. Add the necessary logic to perform
relationship checks between unresolved and resolved classes. If either
class is resolved, we can generally resolve the relationship check by
using the resolved's base list and subclass list.

Note: This is work that would have had to be done anyways. While right
now unresolved classes represent missing classes, in the future they
will instead represent classes whose information is not present on the
current worker shard (truely missing classes will become Bottoms). So,
might as well do it now.

Reviewed By: edwinsmith

Differential Revision: D41069522

fbshipit-source-id: f4bd9638b03fb525eaf9c6948dd8652d4d83ee78