Fix unintentional fall-through in switch-case
Summary:
D35422339 added a code for service throttling. One of its parts adds a `case` clause without a `break`, which causes a fall-through, which in turn causes incorrect code to be assigned.

This was flagged as un-annotated fall-through in https://www.internalfb.com/intern/sandcastle/job/18014399326060987/insights
I thought it's worth fixing.

Reviewed By: roddym

Differential Revision: D41115761

fbshipit-source-id: 9553dc2a2bd74408df785c21e41f2c2510a8b9f2