Make shape key checks monotonic (wrt nullability)
Summary:
It was possible to evade shape key checks provided with coalesce operators and `Shapes::idx` simply by making the shape nullable. This makes the error non-monotonic which is bad because it is surprising to developers and unrelated improvements to completeness can reveal these errors.

This diff extends the checks to nullable shapes and hence making the error monotonic (🤞).

Differential Revision: D41059817

fbshipit-source-id: 52660b25a529a07afc15f851a8db10dadb4f777f