Make SplPriorityQueue<T> conform to Iterator<~T>
Summary:
`SplPriorityQueue::current()` returns "the value or priority (or both) of the current node, depending on the extract flag", according to the PHP docs: https://www.php.net/manual/en/splpriorityqueue.current.php

So most of the time, you can iterate over an `SplPriorityQueue<T>` and see each `T`. But setting the flag will cause it to return something different. That makes it a good candidate for a "like-type", where we offer a type that's "probably" `T` but can be cast to something different if you know you got something else.

Reviewed By: hgoldstein

Differential Revision: D41008708

fbshipit-source-id: 622e20a22dc93c201c1a402003ea4c94d06461d5