Copy down intersected return types to subclasses
Summary:
For explicit targeted pessimisation, we log when a non-fully-enforced return type (e.g. enum `E`) overrides a fully-enforced return type (e.g. `int`), in order to rewrite the return type to an intersected like-type (e.g. `(~E & int)`). Unfortunately we don't do anything about further overrides e.g. if the return type of `E` is further overridden in a subclass, we just pessimise it to `~E`. This leads to hierarchy errors.

Obviously we want to propagate the intersection downwards. Our logging approach makes this a bit tricky. In this diff, whenever a non-fully-enforced return overrides another non-fully-enforced return, we emit a reference back to overridden method, and when processing the log we "close" over these references.

Differential Revision: D41197535

fbshipit-source-id: b528ec62fed4b9f7490d31e5fca2797deb31b5b5