Consider the inner type as the first proposition when doing U <: ~T
Summary:
We've special cased the `T <: (U1 | U2 | ... | Un)` code for like types by extracting the dynamic type and trying on the rest, then doing a `|||` proposition. However, we were checking the `<: dynamic` proposition first, which is almost certainly wrong and most likely slightly slower. We generally expect the majority of cases to be subtyping with the inner type e.g. (`int <: ~int`) and if we see a proposition `string <: ~int`, it's far more informative to the user to report that we expected an `int`, not `dynamic`, especially under implicit pessimisation.

When we get to explicit SDT we should explore reporting "Expected int or dynamic" since then there will be explicit hints to the user in code that they're dealing with a dynamic function, but for implicit mode we're ok with using the current behavior (that only complains about one piece of the union) so long as that is the non-dynamic portion.

Reviewed By: periodic1236

Differential Revision: D41142112

fbshipit-source-id: 186f994a0af1552d8cab38bdbcf9ac1851216f47