Build subclass lists in HHBBC using extern-worker
Summary:
Use extern-worker jobs to build subclass lists (and associated
metadata) in HHBBC. The overall algorithm is quite similar to class
flattening, except for two important differences.

The first is that information flows from children to parents (where as
in class flattening it flowed from parents to children). This doesn't
really change the algorithm for scheduling work. Factor out the
algorithm used in class flattening into a separate function, so that
both passes can utilize it in a generic way.

The second is that the transitive dependency set for each class can be
quite large. A class can have quite a large number of subclasses. This
presents problems because it might not be feasible (or at least
efficient) to load every subclass of a class into an extern-worker
job. Yet those subclasses must be present to calculate the
information.

The solution is two-fold. First, we abandon the strict constraint of
doing everything in one pass. While we try to do that, if a class has
too large a number of subclasses, we'll process some of the
subclasses, then schedule the parent class to be processed in a
subsequent round. This only helps if the subclasses aren't direct
children (we can process intermediate subclasses and reduce the size
of the dependencies).

To help with large number of direct children, we artificially "split"
child lists by adding "split nodes". These are essentially like
classes (though they don't represent an actual class) to the
scheduling algorithm. They exist as a way to process some subset of a
class' child list and provide a place to store the aggregated
results. When processing the actual class, it will take the split
nodes as dependencies and read the intermediate state from them. In
actuality, a class may have a mix of regular classes and split nodes
as dependencies.

As an additional guard on jobs growing too large, we add the ability
to limit any given bucket to a max size. If a bucket exceeds the max
size, it will be split until it's below the size.

As an optimization, we don't process leaf classes for this pass (they
can be used as inputs, but never an output). Leaf classes by
definition have no children, so their metadata can be calculated up
front and not require any changes. We actually calculate all of this
when flattening classes (we assume it has no children). If the class
is actually a leaf, it will be left as is. If not, the information
will be updated when building subclass lists. An additional
optimization is to not process closures (or Closure) since their
subclass information can be predicted ahead of time (and there can be
a lot of these).

Reviewed By: edwinsmith

Differential Revision: D41250803

fbshipit-source-id: 256dec5038d534f664f6fdd4a2220ca0463fe77c