index batch file by file
Summary:
Before this diff, a batch of files was indexed by first indexing the gencode comments, then the declarations, then the xrefs.

We index file by file instead. This simplifies the code and fixes a bug -> we didn't set the proper ownership unit for the gencode comments.

There was no good reason anymore for the previous ordering, it was a vestige of the past.

Reviewed By: pepeiborra

Differential Revision: D41289895

fbshipit-source-id: 8e822968280a4b29b9bb75788804a574ffd1ef77