Check Class User Attributes Outside the Context of the Current Class
Summary:
# Context
User attributes attached to a class were type checked using the same environment used to check individual members of the class. This is not correct, because class bound user attributes are evaluated outside the context of the class. For instance the following would type check:

```
<<SomeAttribute(self::class)>>
class SomeClass {}
```

The `self::class` expression is accepted because we believe we are type checking the user attribute from within the class `SomeClass` when it is clearly outside of it. This intended behavior also occurs when checking modules. Consider:

```
module foo;
internal class InternalFoo {}

<<__Sealed(InternalFoo::class)>>
public trait TFoo {}
```

In this instance we report an error because an internal class is referenced as part of a public trait. However the reference is part of a class bound user attribute, and thus is executed from within the module and not copied to the use of the trait.

# Mitigation

To fix both issues I unset the self type/current_cls in the environment so it is type checked outside the context of the class being checked. This leads to the desired behavior. This change needed to be duplicated in both naming.ml and typing_class.ml. With a larger refactoring I would eliminate the transform done in naming.ml and handle self/parent outside of a class in typing. But I'll go with this more localized change for now.

Reviewed By: hgoldstein

Differential Revision: D41294267

fbshipit-source-id: f8c9e885a7a8724f4d0cb64e9b89baae96ae9ec5