Implement dead unsafe cast removal logic
Summary:
Implements the first version of the logic for removing dead UNSAFE_CASTs.

This diff only adds `hhstc` frontend for testing purposes. The `hh` frontend will follow.

It only handles the simplest case where an `UNSAFE_CAST` will be removed and only if the input of the `UNSAFE_CAST` has a type that makes the cast redundant, e.g.,

```
function f(int $i): void {
  UNSAFE_CAST<string, int>($i); // Redundant since $i is already an int
}
```

The more complicated case where it is redundant because the UNSAFE_CAST wasn't needed at all is not covered, e.g.,

```
function f(string $i): void {
   // Redundant because the result of the cast is not used
  UNSAFE_CAST<string, int>($i);
}
```
or
```
function f(arraykey $i): arraykey {
   // Redundant because the return type does not need the return expression to be a string
  return UNSAFE_CAST<arraykey, string>($i);
}
```

Because we don't yet have a way of detecting when we are doing the transformation inside an expression that can capture (e.g. `2 * UNSAFE_CAST<...,...>(...)`) we add parantheses around the replaced expression which is redundant most of the time.

Another limitation of this implementation is that it doesn't preserve comments of the argument to the `UNSAFE_CAST` which might be an `HH_FIXME`. There's a test case that captures this behaviour. This will be addressed in later iterations of the tool.

Reviewed By: enetsee

Differential Revision: D41061011

fbshipit-source-id: 36841d55cb231eab1575bb8537912923d5aa41e1