Allow Internal Type Aliases to Refer to Internal Types
Summary:
# Context
When the ability to mark type aliases as internal was added, we forgot to call Typing_env.set_internal when checking a type alias. This leads to erroring when an internal type alias refers to another internal type. For instance:

```
internal type MyInt = int;
internal type MyVecInt = vec<MyInt>; // Incorrectly reports error
```
We report an error for using the internal type `MyInt` as part of the type alias `MyVecInt`. Since `MyVecInt` is also internal, this should be allowed.

# Mitigation
Properly set internal in the environment when checking typedefs

Reviewed By: hgoldstein

Differential Revision: D41313015

fbshipit-source-id: d013a21ec30135f2dbe978c5368c0032925c3fb3