Make FilesystemIterator::current() return ~SplFileInfo
Summary:
By default, `FilesystemIterator::current()` returns an `SplFileInfo`, but depending on flags it could return a `string` or itself.

https://www.internalfb.com/code/fbsource/[06fa37ff85dd96d9992b473c7b1821d783049e93]/fbcode/hphp/system/php/spl/iterators/FilesystemIterator.php?lines=63-71

`DirectoryIterator::current()` always returns itself in its implementation...

https://www.internalfb.com/code/fbsource/[06fa37ff85dd96d9992b473c7b1821d783049e93]/fbcode/hphp/system/php/spl/iterators/DirectoryIterator.php?lines=59-61

but because `FilesystemIterator extends DirectoryIterator`, we have to acknowledge that a given `DirectoryIterator` might actually be a `FilesystemIterator` and may therefore return a `string`.

So make `DirectoryIterator` and `FilesystemIterator` both implement `Iterator<~SplFileInfo>`. `DirectoryIterator` extends `SplFileInfo`. So when it returns itself, it is returning an object of the correct type.

Reviewed By: hgoldstein

Differential Revision: D41008379

fbshipit-source-id: 9bfb21f447d7bd5a3edfb92cef72022537b6c550