Explicitly pass the exit block to irgen::checkType()
Summary:
Kill IRBuilder::m_guardFailBlock. It was used only by irgen::checkType(), which
was always called close to the place where the IRBuilder state was configured,
making it easy to just pass the exit block explicitly.

This makes it easy to track where the exit block is coming from and also makes
irgen::checkType() less tied to the concept of guarding.

Reviewed By: ricklavoie, ottoni

Differential Revision: D41132599

fbshipit-source-id: a5821fb8192322cf1b90a1e3eea642a0a17f636f