Add hhi for classname_from_string_unsafe
Summary: {D40284561 (https://github.com/facebook/hhvm/commit/bc7260465933d884cb126dbd2bb5e31609465c37)} added a function to backdoor strings to be treated as classnames (LazyClass) by the runtime. This is needed to address sources of module violations present in the codebase that are benign. Given the HHVM change landed more than a month ago it is safe to expose this function via hhi's.

Reviewed By: oulgen

Differential Revision: D41324884

fbshipit-source-id: 3177aa8d99ded556bd956d4b7ee577eb0b5650d4