Hoist include path logic out of printer
Summary:
We do a bunch of work on `IncludePaths` in the bytecode printer. We don't get any of the benefit of this in HackC-Translator. If we hoist all this work into the emitter, we can make the printer more pure, and allow HackC-Translator to share the logic.

This diff separates resolution of roots using include_roots/filepath, with actual existence checks.

Changes to bytecode are just the omission of empty `include{}` segments. Since we do processing before bytecode_printer, we realize paths don't exist before we print the opening `include {`

Reviewed By: aorenste

Differential Revision: D41240665

fbshipit-source-id: a13c9d84acadc0533dc08da5e93b39e0ef9b83f2