simplify Naming_provider resolution of Pos_or_decl
Summary:
This diff concerns Naming_provider.get_{module,const,fun,type}_full_pos.
* Input: `FileInfo.name_type` for a symbol, `Relative_path.t` where the symbol is defined, `string` name of symbol
* Output: `Pos.t` which includes filename, line, column information.

In hh_server this is implemented by parsing `fn` into the AstProvider cache, walking it to find the symbol, and getting the full position out of the AST.

In the re-architecture we'd honestly rather not do that. We're relying for now that the decls provided by the decl service include at least Pos_or_decl.t that includes line and column information, e.g. for a typedef we look at `Typing_defs.typing_type:td_pos` field, and for a fun we look at `Typing_defs.fun_elt:fe_pos`. What we do with that is substitute in the correct Relative_path, and we've got ourselves the Pos.t.

Except, what we're doing is a bit weird. We are already given a `Relative_path.t` but we discard it. Instead we make a second query to the naming table to learn the filename associated with the symbol, and use that. This behavior was introduced in D27325491 (https://github.com/facebook/hhvm/commit/717c257be6e40c3712a93185381af47bdbd7f943) (March 2021)

With this diff I switched to just using the filename we're given in the first place.

Reviewed By: CatherineGasnier

Differential Revision: D41312279

fbshipit-source-id: ea9812414b7f15f59504776a3865190b4550292e