Don't raise false positives for HackLint5624 in type refinements [1/2]: exact bounds
Summary:
When a generic parameter is used only in the type refinement, e.g.

  function f<TNum as num>(): Box with { type T = TNum }

the linter previously erred because every type of bound (equal, as, super)
was treated the same.  Fix this by looking inside the refinements,

For equal bounds, treat the generic as both positive and negative w.r.t. variance, which is consistent with where-clauses.
This is so that a generic which is bounded, such as `TNum as num` in the example,
doesn't get a single classification (either positive or negative) w.r.t variance,
which in turn raises the false positive in the linter; i.e.:

```
The generic parameter `T0` is redundant because it only appears in a contravariant (input) position. Consider replacing uses of generic parameter with `shape(...)` or specifying `<<__Explicit>>` on the generic parameter (Lint[5624])

File "redundant_generics_type-refinements.php", line 17, characters 34-35:
The generic parameter `T1` is redundant because it only appears in a covariant (output) position with useless `as` bound. Consider replacing uses of generic parameter with `nothing` or specifying `<<__Explicit>>` on the generic parameter (Lint[5624])
```

Reviewed By: mpu

Differential Revision: D41330243

fbshipit-source-id: b38c6e12d17b72017035d600942e7e0a4ebc7441