Do not unnecessarily guard moved or DecRef'd inputs
Summary:
Do not guard BStackN -- these inputs are used only by CreateCl to move closure
use variables from stack to the closure, or to DecRef inputs of final minstrs.
Those are usually the final DecRef uses, so even DataTypeCountness is not going
to give us anything.

Do not guard top of the stack of SetM, this stack input is just being copied to
the destination, followed by a very likely PopC, which will DecRef it.

Reviewed By: ottoni, oulgen

Differential Revision: D41362823

fbshipit-source-id: be597eb86569b96469ba5183926c27b6efead86d