Optimize edges to simple Check{TypeMem,Loc,Stk,MBase,InitMem,MROProp} blocks
Summary:
If the next or taken edge of a block points to a simple block with a single
Check* branch instruction, try to use the stateOutNext or stateOutTaken to
statically resolve the branch. If possible, bypass the simple block.

This conflicts with creation of phis, as we may no longer need a phi at the
original place and instead may need a phi at a different place that gets
additional input edges. Solve this problem by running optimization pass in two
stages, first without phis, and only if we did not perform any CFG changes in
that pass, replace loads with phis. This eliminates some unnecessary phis in
pre-existing situations, where FJmpNext or FJmpTaken made CFG changes.

The main effect of this diff is that N sequences of linear checks for M guards
previously performing O(N*M) checks total will be rewritten into a tree of
checks with the maximum number of checks performed being O(N+M).

Reviewed By: ottoni

Differential Revision: D41327955

fbshipit-source-id: 8959ab4ceb67b52ec4c64819e5de452c790ec9cf