Make header take precendence over rocket
Summary:
When WWW sets an outbound thrift_tfmk HTTP header, the intermediate infra will also automatically generate a Rocket-based frameworkMetadata payload (with a newly-generated RequestId). This is taking precedence in C++'s thrift server. This should not be the case.

Attempt to read the header value first. This also requires plumbing a return value through the hook.

Reviewed By: bsnsk

Differential Revision: D41288359

fbshipit-source-id: d2b0228bbd0c92c8688324da86a59be2df1faa74