Allow global const decls in HHIs to omit the value
Summary:
Global consts cannot be abstract and therefore have to declare something for their value.

In HHI files, however, we are only going to use the declaration to know that the name exists and for its type (which is either derived from the typehint or inferred from the provided value). If a type has been specified, then the value is redundant. Either way, the value in the HHI is not used in codegen and is typically just a placeholder. In some cases we cannot even specify a value with the correct type because the builtin constant is magic.

Therefore we should not require a value for global constant declarations when in HHI mode.

Future changes will likely remove the value part entirely from HHI files and forbid them from being specified to avoid confusion. We may do the same for concrete class constants.

Reviewed By: hgoldstein

Differential Revision: D41232315

fbshipit-source-id: 0fd2a7f776aecb103068cee8f73bf66ef04cc01a