use list of events to handle watchman subscriptions when out of order
Summary:
this diff attempts to tackle S301212 by holding onto a stack of state
transitions from Watchman subscription events. In the previous implementation,
we kept a boolean flag noting if we were in an hg.transaction or hg.update
state.

The SEV notes a strange case where, upon reserving an ondemand, we sometimes
see incomplete and out of order state changes that look something like

transaction/leave
transaction/leave
transaction/leave
transaction/enter
transaction/enter

My original idea was to keep a stack of events, where we push timestamps for
ENTER and pop for LEAVE. If we get a LEAVE and the stack is empty, we hold it
in a buffer stack. Then if an ENTER comes, we try to clear our buffer stack
FIRST.

This becomes uniquely challenging when dealing with unpaired ENTER/LEAVE
events. Which stack do you clear? Under what circumstance? ljw1004 noted a cleaner
solution, using a single list of events, walking the list to pop paired events
when an event comes in, and expiring events on staleness.

Reviewed By: ljw1004

Differential Revision: D41282155

fbshipit-source-id: e24888d66fd85ce50e510332e7796b4b33ffab91