Lower Array constants.
Summary:
Move constant lowering into the lowering phase.

Tweak the code so only Strings and Arrays are considered "complex constants" (the others are emitted inline).

Teach the code how to emit TypedValue so array constants can be emitted.

Also use a simple call for VerifyParamTypeTS instead of an inlined branch.

Reviewed By: davidpichardie

Differential Revision: D41259168

fbshipit-source-id: 3dde77b58e057159368a32fbad64fd12e09af89a